+++ lib/volt/router/routes.rb
@@ -5,38 +5,38 @@ module Volt
   # a url to params, and params to url.
   # routes do
   #   client "/about", _view: 'about'
-  #   client "/blog/{id}/edit", _view: 'blog/edit', _action: 'edit'
-  #   client "/blog/{id}", _view: 'blog/show', _action: 'show'
-  #   client "/blog", _view: 'blog'
-  #   client "/blog/new", _view: 'blog/new', _action: 'new'
-  #   client "/cool/{_name}", _view: 'cool'
+  #   client "/blog/{{ id }}/edit", view: 'blog/edit', action: 'edit'
+  #   client "/blog/{{ id }}", view: 'blog/show', action: 'show'
+  #   client "/blog", view: 'blog'
+  #   client "/blog/new", view: 'blog/new', action: 'new'
+  #   client "/cool/{{ name }}", view: 'cool'
   # end
   #
   # Using the routes above, we would generate the following:
   #
   # @direct_routes = {
-  #   '/about' => {_view: 'about'},
-  #   '/blog' => {_view: 'blog'}
-  #   '/blog/new' => {_view: 'blog/new', _action: 'new'}
+  #   '/about' => {view: 'about'},
+  #   '/blog' => {view: 'blog'}
+  #   '/blog/new' => {view: 'blog/new', action: 'new'}
   # }
   #
   # -- nil represents a terminal
-  # -- * represents any match
+  # -- * represents any match (in that section (between / and /))
   # -- a number for a parameter means use the value in that number section
   #
   # @indirect_routes = {
   #     '*' => {
   #       'edit' => {
-  #         nil => {id: 1, _view: 'blog/edit', _action: 'edit'}
+  #         nil => {id: 1, view: 'blog/edit', action: 'edit'}
   #       }
-  #       nil => {id: 1, _view: 'blog/show', _action: 'show'}
+  #       nil => {id: 1, view: 'blog/show', action: 'show'}
   #     }
   #   }
   # }
   #
   # Match for params
   # @param_matches = [
-  #   {id: nil, _view: 'blog/edit', _action: 'edit'} => Proc.new {|params| "/blog/#{params.id}/edit", params.reject {|k,v| k == :id }}
+  #   {id: nil, view: 'blog/edit', action: 'edit'} => Proc.new {|params| "/blog/#{params.id}/edit", params.reject {|k,v| k == :id }}
   # ]
   class Routes
     def initialize
@@ -68,7 +68,7 @@ module Volt
     end
 
     # Add server side routes
-    
+
     def get(path, params)
       create_route(:get, path, params)
     end
@@ -190,17 +190,29 @@ module Volt
       if part.nil?
         if node[part]
           # We found a match, replace the bindings and return
-          # TODO: Handvle nested
+          # TODO: Handle nested
           setup_bindings_in_params(original_parts, node[part])
         else
           false
         end
-      elsif (new_node = node[part])
-        # Direct match for section, continue
-        match_path(original_parts, parts, new_node)
-      elsif (new_node = node['*'])
-        # Match on binding section
-        match_path(original_parts, parts, new_node)
+      else
+        if (new_node = node[part])
+          # Direct match for section, continue
+          result = match_path(original_parts, parts, new_node)
+          return result if result
+        end
+        if (new_node = node['*'])
+          # Match on binding single section
+          result = match_path(original_parts, parts, new_node)
+          return result if result
+        end
+        if ((params = node['**']) && params && (params = params[nil]))
+          # Match on binding multiple sections
+          result = setup_bindings_in_params(original_parts, params)
+          return result if result
+        end
+
+        return false
       end
     end
 
@@ -214,6 +226,10 @@ module Volt
         if value.is_a?(Fixnum)
           # Lookup the param's value in the original url parts
           params[key] = original_parts[value]
+        elsif value.is_a?(Range)
+          # When doing multiple section bindings, we lookup the parts as a range
+          # then join them with /
+          params[key] = original_parts[value].join('/')
         end
       end
 
@@ -231,10 +247,24 @@ module Volt
       parts.each_with_index do |part, index|
         if has_binding?(part)
           # Strip off {{ and }}
-          params[part[2...-2].strip.to_sym] = index
+          section_matcher, multipart = binding_extract(part)
 
-          # Set the part to be '*' (anything matcher)
-          part = '*'
+          if multipart
+            # Match anything for the rest of the url (multiple sections)
+
+            # check that the splat is at the end of the url
+            if index != (parts.size-1)
+              raise "The splat (*) operator can only be used at the end of a url"
+            end
+
+            part = '**'
+            index = (index..-1)
+          else
+            # Match anything in a single section, set the part to be '*'
+            part = '*'
+          end
+
+          params[section_matcher] = index
         end
 
         node = (node[part] ||= {})
@@ -251,7 +281,8 @@ module Volt
         if has_binding?(part)
           # Setup a nil param that can match anything, but gets
           # assigned into the url
-          params[part[2...-2].strip.to_sym] = nil
+          section_matcher, _ = binding_extract(part)
+          params[section_matcher] = nil
         end
       end
 
@@ -270,7 +301,7 @@ module Volt
         url = parts.map do |part|
           val = if has_binding?(part)
                   # Get the
-                  binding = part[2...-2].strip.to_sym
+                  binding, _ = binding_extract(part)
                   input_params.delete(binding)
                 else
                   part
@@ -330,6 +361,23 @@ module Volt
     #Append an id to a given path
     def path_with_id(base_path)
       base_path + '/{{ id  }}'
+    end
+
+    # @param: The binding part of the url eg: "{{ binding }} or {{ *binding }}"
+    # @return: a symbol for the binding variable eg: :binding, and a boolean for
+    #          if the binding matches multiple sections.
+    def binding_extract(part)
+      section_matcher = part[2...-2].strip
+
+      multipart_matcher = false
+
+      if section_matcher[0] == '*'
+        # Match anything for the rest of the url (multiple sections)
+        section_matcher = section_matcher[1..-1]
+        multipart_matcher = true
+      end
+
+      return section_matcher.to_sym, multipart_matcher
     end
   end
 end