+++ lib/volt/server/rack/sprockets_helpers_setup.rb
@@ -38,6 +38,7 @@ module Volt
         # We "freedom-patch" sprockets-helpers asset_path method to
         # automatically link assets.
         define_method(:asset_path) do |source, options = {}|
+          # puts "AP: #{source.inspect}"
           relative_path = source =~ /^[.][.]\//
           if relative_path
             component_root = logical_path.gsub(/\/[^\/]+$/, '')
@@ -104,3 +105,32 @@ module Volt
     end
   end
 end
+
+
+# module Sprockets
+#   class Context
+#     # We have to reinclude this because of some ```require``` thread saftey issues.
+#     def find_asset_path(uri, source, options = {})
+#       if Helpers.manifest && options[:manifest] != false
+#         manifest_path = Helpers.manifest.assets[uri.path]
+#         return Helpers::ManifestPath.new(uri, manifest_path, options) if manifest_path
+#       end
+
+#       if Sprockets::Helpers.are_using_sprockets_3
+#         resolved = assets_environment.resolve(uri.path)
+
+#         if resolved
+#           return Helpers::AssetPath.new(uri, assets_environment[uri.path], options)
+#         else
+#           return Helpers::FilePath.new(uri, options)
+#         end
+#       else
+#         assets_environment.resolve(uri.path) do |path|
+#           return Helpers::AssetPath.new(uri, assets_environment[path], options)
+#         end
+
+#         return Helpers::FilePath.new(uri, options)
+#       end
+#     end
+#   end
+# end