+++ spec/router/routes_spec.rb
@@ -92,6 +92,10 @@ describe Volt::Routes do
       put '/people', controller: 'people', action: 'update'
       patch '/people/1', controller: 'people', action: 'update'
       delete '/people/2', controller: 'people', action: 'destroy'
+      client '/pics/new/{{ new_id }}/two', controller: 'pics', action: 'new'
+      client '/pics/{{ pic_id }}/view/one', controller: 'pics', action: 'view'
+      client '/pics/{{ *pic_id }}', controller: 'pics', action: 'show'
+      get '/wide_match/{{ *path_parts }}', controller: 'wide_match', action: 'show'
     end
 
     params = @routes.url_to_params('/blog')
@@ -122,7 +126,7 @@ describe Volt::Routes do
     expect(params).to eq(controller: 'articles', action: 'index')
 
     params = @routes.url_to_params(:post, '/articles')
-    expect(params).to be_nil
+    expect(params).to eq(false)
 
     params = @routes.url_to_params(:post, '/comments')
     expect(params).to eq(controller: 'comments', action: 'create')
@@ -141,6 +145,26 @@ describe Volt::Routes do
 
     params = @routes.url_to_params(:put, '/articles/2/comments/9')
     expect(params).to eq(controller: 'comments', action: 'update', articles_id: '2', id: '9')
+
+    params = @routes.url_to_params('/pics/new/view/one')
+    expect(params).to eq({controller: "pics", action: "view", pic_id: "new"})
+
+    params = @routes.url_to_params('/pics/new/view/two')
+    expect(params).to eq({controller: "pics", action: "new", new_id: "view"})
+
+    params = @routes.url_to_params('/pics/new/view/three')
+    expect(params).to eq({controller: "pics", action: "show", pic_id: "new/view/three"})
+
+    params = @routes.url_to_params(:get, '/wide_match/some/image/path/file.png')
+    expect(params).to eq({controller: "wide_match", action: "show", path_parts: "some/image/path/file.png"})
+  end
+
+  it 'should raise an error if the splat match isn\'t at the end of the url' do
+    expect do
+      routes do
+      client '/blog/{{ *slug }}/after', view: 'blog'
+      end
+    end.to raise_error('The splat (*) operator can only be used at the end of a url')
   end
 
   it 'should go from params to url' do
@@ -279,16 +303,16 @@ describe Volt::Routes do
     params = @routes.url_to_params(:get, '/api/v1/articles/1')
     expect(params).to eq(controller: 'articles', action: 'show', id: '1')
 
-    expect(@routes.url_to_params(:post, '/api/v1/articles')).to be_nil
-    expect(@routes.url_to_params(:put, '/api/v1/articles/1')).to be_nil
-    expect(@routes.url_to_params(:delete, '/api/v1/articles/1')).to be_nil
+    expect(@routes.url_to_params(:post, '/api/v1/articles')).to eq(false)
+    expect(@routes.url_to_params(:put, '/api/v1/articles/1')).to eq(false)
+    expect(@routes.url_to_params(:delete, '/api/v1/articles/1')).to eq(false)
 
     routes do
       rest '/api/v1/articles', controller: 'articles', only: [:update, :destroy, :create]
     end
 
-    expect(@routes.url_to_params(:get, '/api/v1/articles')).to be_nil
-    expect(@routes.url_to_params(:get, '/api/v1/articles/1')).to be_nil
+    expect(@routes.url_to_params(:get, '/api/v1/articles')).to eq(false)
+    expect(@routes.url_to_params(:get, '/api/v1/articles/1')).to eq(false)
 
     params = @routes.url_to_params(:post, '/api/v1/articles')
     expect(params).to eq(controller: 'articles', action: 'create')
@@ -311,8 +335,8 @@ describe Volt::Routes do
     params = @routes.url_to_params(:get, '/api/v1/articles/1')
     expect(params).to eq(controller: 'articles', action: 'show', id: '1')
 
-    expect(@routes.url_to_params(:post, '/api/v1/articles')).to be_nil
-    expect(@routes.url_to_params(:put, '/api/v1/articles/1')).to be_nil
-    expect(@routes.url_to_params(:delete, '/api/v1/articles/1')).to be_nil
+    expect(@routes.url_to_params(:post, '/api/v1/articles')).to eq(false)
+    expect(@routes.url_to_params(:put, '/api/v1/articles/1')).to eq(false)
+    expect(@routes.url_to_params(:delete, '/api/v1/articles/1')).to eq(false)
   end
 end